+++ R/nesting.R
@@ -21,9 +21,9 @@
 #'   color = c("blue", "red", "green")
 #' )
 #'
-#' funnel %>%
-#'   e_charts() %>%
-#'   e_funnel(value, stage) %>%
+#' funnel |>
+#'   e_charts() |>
+#'   e_funnel(value, stage) |>
 #'   e_add("itemStyle", color)
 #'
 #' # Heatmap can take nested label
@@ -34,19 +34,19 @@
 #'   y = sample(v, 300, replace = TRUE),
 #'   z = rnorm(300, 10, 1),
 #'   stringsAsFactors = FALSE
-#' ) %>%
-#'   dplyr::group_by(x, y) %>%
-#'   dplyr::summarise(z = sum(z)) %>%
-#'   dplyr::ungroup() %>%
+#' ) |>
+#'   dplyr::group_by(x, y) |>
+#'   dplyr::summarise(z = sum(z)) |>
+#'   dplyr::ungroup() |>
 #'   dplyr::mutate(
 #'     show = TRUE,
 #'     fontStyle = round(runif(dplyr::n(), 5, 12))
 #'   )
 #'
-#' matrix %>%
-#'   e_charts(x) %>%
-#'   e_heatmap(y, z) %>%
-#'   e_visual_map(z) %>%
+#' matrix |>
+#'   e_charts(x) |>
+#'   e_heatmap(y, z) |>
+#'   e_visual_map(z) |>
 #'   e_add(
 #'     "label",
 #'     show,
@@ -59,7 +59,7 @@ e_add <- function(e, param, ...) {
   }
 
   for (i in seq_along(e$x$data)) {
-    data <- e$x$data[[i]] %>%
+    data <- e$x$data[[i]] |>
       dplyr::select(...)
 
     data <- apply(data, 1, as.list)